// src/app/model/arrestee.model.ts
import {LawOffenseModel} from './law-offense.model';
import {DisciplineViolationModel} from './discipline-violation.model';
import {ArrestSettlementDecisionModel} from './arrest-settlement-decision.model';

export interface ArresteeModel {
  id: number;
  arrestDetentionInfoId: number;
  arrestType: string;
  caseId: string;
  caseName: string;
  defendantId: string;
  defendantName: string;
  arrestDate: Date;
  procuracyHandlingDate: Date;
  arrestReason: string;
  arrestViolation: string;
  fullName: string;
  dateOfBirth: Date;
  yearOfBirth: number;
  job: string;
  workplace: string;
  address: string;
  isDead: boolean;
  deathDate: Date;
  causeOfDeathId: string;
  causeOfDeathName: string;
  escaped: boolean;
  escapingDate: Date;
  reasonForEscaping: string;
  recaptureDate: Date;
  moveToAnotherPlace: boolean;
  moveToAnotherPlaceDate: Date;
  arriveFromAnotherPlace: boolean;
  arriveFromAnotherPlaceDate: Date;
  reason: string;
  createdBy: string;
  createdAt: any;
  updatedBy: string;
  updatedAt: any;
  status: number;
  arresteeYear: number;
  arresteeMonth: number;
  arresteeDay: number;
  noiChuyenDi: string;
  noiChuyenDen: string;
  detentionPlace: string;
  lawOffenses: LawOffenseModel[];
  disciplineViolations: DisciplineViolationModel[];
  settlementDecisions: ArrestSettlementDecisionModel[];
  cccd: number;
}